llvc: split MainWindow utility logic into helper translation unit
